backend/requirements.txt
flask
flask-cors
tensorflow
Pillow
gunicorn
numpy
